requirements.txt
# requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
transformers>=4.20.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.20.0
pyyaml>=6.0
tqdm>=4.64.0
plotly>=5.0.0
gymnasium>=0.28.0
flask>=2.0.0
flask-cors>=3.0.0
redis>=4.0.0
schedule>=1.2.0
pyarrow>=10.0.0